requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.4.0
bcrypt==3.2.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6       
fastapi==0.115.0      
greenlet==3.1.1       
h11==0.14.0
httptools==0.6.1      
idna==3.10
Mako==1.3.5
MarkupSafe==2.1.5     
passlib==1.7.4        
psycopg2==2.9.9       
pycparser==2.22       
pydantic==2.9.2       
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlmodel==0.0.22
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.0.1